refactor(web-todomvc): migrate to new multi-store API
Replace LiveStoreProvider with StoreRegistryProvider pattern:
- Create centralized store.ts with useAppStore hook
- Add ErrorBoundary and Suspense above StoreRegistryProvider
- Update components to use useAppStore() instead of useStore()
